wip: try to handle responses from Verify
Added a script to initialize the compliance tool and an example SAML
response.

Needed to do a bit of fiddling - it turns out passport-saml needs
body-parser to be installed for it to handle post responses.

Now passport-saml *is* handling the SAML responses, but I'm not
confident it's doing so securely. It's also not completely happy yet for
some reason I have yet to identify.